Updated cart service with various operations and kafka producer calls

+++ cart-service/src/main/java/com/tanzeem/cart_service/entity/UserCart.java
@@ -29,7 +29,7 @@ public class UserCart {
 	private String status;										//ACTIVE, ORDERED, ABANDONED
 	
 	
-	private BigDecimal totalAmount;
+	private double totalAmount;
 	
 	private LocalDateTime createdAt;
 	private LocalDateTime updatedAt;
